Wire trained Victorian stylizer as a selectable backend
- pipeline: stylize_document(..., backend="llm"|"t5"); "t5" routes to the
  trained Seq2SeqStylizer (models/stylizer-victorian) and falls back to the
  LLM stylizer if the model dir is absent. Per-backend lazy cache.
- app: reader Read tab gains an "Engine" dropdown — "Fast T5 (Victorian)"
  (offline, register baked in) vs "LLM (rich)" (prompted, all presets).
- finetune_stylizer: use bf16 (flan-t5 is unstable in fp16).

+++ training/finetune_stylizer.py
@@ -105,7 +105,7 @@ def main():
         generation_max_length=args.max_length,
         group_by_length=True,
         logging_steps=50,
-        fp16=torch.cuda.is_available(),
+        bf16=torch.cuda.is_available(),   # T5 is unstable in fp16 (NaN); bf16 is safe on Ada
         report_to=[],
     )
     trainer = Seq2SeqTrainer(
